security_pattern_miner/requirements.txt
pydantic
requests
# python-woc
jsonlines
tqdm
GitPython
pyYAML